dostumqunduz/requirements.txt
django
djangorestframework
Pillow
django-filter
markdown